+++ src/String.php
@@ -6,9 +6,10 @@ namespace Phlib\String;
 
 function ellipsis(string $string, int $maxLength, string $ellipsis = '...'): string
 {
+    $ellipsisLength = mb_strlen($ellipsis);
     $options = [
         'options' => [
-            'min_range' => 1,
+            'min_range' => 1 + $ellipsisLength,
             'max_range' => PHP_INT_MAX,
         ],
     ];
@@ -17,8 +18,8 @@ function ellipsis(string $string, int $maxLength, string $ellipsis = '...'): str
         throw new \InvalidArgumentException("Cannot use value provided as maxlength '{$maxLength}'");
     }
 
-    if (strlen($string) > $maxLength) {
-        return substr($string, 0, $maxLength - strlen($ellipsis)) . $ellipsis;
+    if (mb_strlen($string) > $maxLength) {
+        return mb_substr($string, 0, $maxLength - $ellipsisLength) . $ellipsis;
     }
     return $string;
 }